+++ src/org/xresloader/core/data/dst/DataDstJson.java
@@ -243,7 +243,7 @@ public class DataDstJson extends DataDstJava {
                 final String numberAsString = JSONObject.numberToString((Number) value);
                 if (NUMBER_PATTERN.matcher(numberAsString).matches()) {
                     // IEEE 754
-                    if (value instanceof Long
+                    if (SchemeConf.getInstance().getJsonOptions().enableLargeNumberAsString && value instanceof Long
                             && ((long) value > ((1L << 53) - 1) || (long) value < -((1L << 53) - 1))) {
                         JSONObject.quote(numberAsString, writer);
                     } else {