+++ src/org/xresloader/core/scheme/SchemeDataSourceBase.java
@@ -60,6 +60,8 @@ public abstract class SchemeDataSourceBase implements SchemeDataSourceImpl {
             SchemeConf.getInstance().getKey().buildKeyWordRegexPrefixRule(datas.get(2));
         } else if (key.equalsIgnoreCase("Encoding")) {
             SchemeConf.getInstance().getKey().setEncoding(datas.get(0));
+        } else if (key.equalsIgnoreCase("JsonCfg-LargeNumberAsString")) {
+            SchemeConf.getInstance().setJsonOptionEnableLargeNumberAsString(datas.get(0));
         } else if (key.equalsIgnoreCase("UeCfg-UProperty")) {
             SchemeConf.getInstance().setUEOptions(datas.get(0), datas.get(1), datas.get(2));
         } else if (key.equalsIgnoreCase("UeCfg-CodeOutput")) {